+++ base/include/event_json_parser.h
@@ -35,7 +35,7 @@ constexpr uint8_t INVALID_EVENT_TYPE = 0;
 constexpr uint8_t DEFAULT_PRIVACY = 4;
 constexpr uint8_t DEFAULT_PERSERVE_VAL = 1;
 constexpr uint8_t DEFAULT_COLLECT_VAL = 0;
-constexpr int32_t DEFAULT_REPORT_INTERVAL = 0;
+constexpr int16_t DEFAULT_REPORT_INTERVAL = 0;
 
 struct KeyConfig {
     uint8_t type : 3;
@@ -53,7 +53,7 @@ struct BaseInfo {
     std::string level;
     std::string tag;
     PARAM_INFO_MAP_PTR disallowParams;
-    int32_t reportInterval = DEFAULT_REPORT_INTERVAL;
+    int16_t reportInterval = DEFAULT_REPORT_INTERVAL;
 };
 
 using NAME_INFO_MAP = std::unordered_map<std::string, BaseInfo>;
